+++ Contexts/MainAppContext.cs
@@ -1,4 +1,4 @@
-﻿using AonFreelancing.Models;
+﻿using AonFreelancing.Models.Entities;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 using static System.Net.WebRequestMethods;
@@ -39,12 +39,13 @@ namespace AonFreelancing.Contexts
                                     .HasPrincipalKey<TempUser>(nameof(TempUser.PhoneNumber));
 
             builder.Entity<Project>().ToTable("Projects", tb => tb.HasCheckConstraint("CK_PRICE_TYPE", "[PriceType] IN ('Fixed', 'PerHour')"));
-            
             builder.Entity<Project>()
                 .ToTable("Projects", tb => tb.HasCheckConstraint("CK_QUALIFICATION_NAME", "[QualificationName] IN ('uiux', 'frontend', 'mobile', 'backend', 'fullstack')"));
             builder.Entity<Project>().ToTable("Projects", tb => tb.HasCheckConstraint("CK_STATUS", "[Status] IN ('Available', 'Closed')"))
                 .Property(p=>p.Status).HasDefaultValue("Available");
-            
+            builder.Entity<Project>().Property(p => p.PriceType).HasDefaultValue("Fixed");
+
+
             base.OnModelCreating(builder);
         }
     }
