+++ Controllers/Mobile/v1/ProjectsController.cs
@@ -1,6 +1,6 @@
 ﻿using AonFreelancing.Contexts;
-using AonFreelancing.Models;
 using AonFreelancing.Models.DTOs;
+using AonFreelancing.Models.Entities;
 using AonFreelancing.Services;
 using AonFreelancing.Utilities;
 using Microsoft.AspNetCore.Authorization;
