+++ Models/Entities/Freelancer.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AonFreelancing.Models
+namespace AonFreelancing.Models.Entities
 {
     [Table("Freelancers")]
     public class Freelancer : User
