+++ Models/Entities/TempUser.cs
@@ -3,16 +3,16 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.AspNetCore.Identity;
 
-namespace AonFreelancing.Models;
+namespace AonFreelancing.Models.Entities;
 
 [Table("TempUser")]
-public class TempUser 
-{ 
+public class TempUser
+{
     [Key]
     public long Id { get; set; }
     [Required]
     public string PhoneNumber { get; set; }
     public bool PhoneNumberConfirmed { get; set; }
-    [Required, AllowedValues("FREELANCER", "CLIENT")] 
+    [Required, AllowedValues("FREELANCER", "CLIENT")]
     public string UserType { get; set; }
 }
