+++ Models/Entities/User.cs
@@ -9,7 +9,7 @@ using AonFreelancing.Interfaces;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 
-namespace AonFreelancing.Models
+namespace AonFreelancing.Models.Entities
 {
     //Replaced with 'builder.Entity<User>().HasIndex(u => u.PhoneNumber).IsUnique();' in OnModelCreate
     //[Index(nameof(PhoneNumber), IsUnique = true)]
@@ -17,6 +17,6 @@ namespace AonFreelancing.Models
     public class User : IdentityUser<long>
     {
         public string Name { get; set; }
-        public string? About {  get; set; }
+        public string? About { get; set; }
     }
 }
